Renamed Project from MenuAnimado1 to MenuAnimation
