--- MenuAnimado1/Properties/AssemblyInfo.cs
@@ -1,55 +1,0 @@
-﻿using System.Reflection;
-using System.Resources;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-using System.Windows;
-
-// As informações gerais sobre um assembly são controladas por
-// conjunto de atributos. Altere estes valores de atributo para modificar as informações
-// associada a um assembly.
-[assembly: AssemblyTitle("MenuAnimado1")]
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("MenuAnimado1")]
-[assembly: AssemblyCopyright("Copyright ©  2017")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-// Definir ComVisible como false torna os tipos neste assembly invisíveis
-// para componentes COM. Caso precise acessar um tipo neste assembly de
-// COM, defina o atributo ComVisible como true nesse tipo.
-[assembly: ComVisible(false)]
-
-//Para começar a compilar aplicativos localizáveis, configure
-//<UICulture>CultureYouAreCodingWith</UICulture> no seu arquivo .csproj
-//dentro de um <Grupo de Propriedade>.  Por exemplo, se você está usando o idioma inglês
-//nos seus arquivos de origem, configure o <UICulture> para en-US.  Em seguida, descomente
-//o atributo NeutralResourceLanguage abaixo.  Atualize o "en-US" na
-//linha abaixo para coincidir com a configuração do UICulture no arquivo do projeto.
-
-//[assembly: NeutralResourcesLanguage("en-US", UltimateResourceFallbackLocation.Satellite)]
-
-
-[assembly: ThemeInfo(
-    ResourceDictionaryLocation.None, //onde os dicionários de recursos de temas específicos estão localizados
-                                     //(usado se algum recurso não for encontrado na página,
-                                     // ou dicionários de recursos do aplicativo)
-    ResourceDictionaryLocation.SourceAssembly //onde o dicionário de recursos genéricos está localizado
-                                              //(usado se algum recurso não for encontrado na página,
-                                              // app, ou qualquer outro dicionário de recursos de tema específico)
-)]
-
-
-// As informações da versão de um assembly consistem nos quatro valores a seguir:
-//
-//      Versão Principal
-//      Versão Secundária 
-//      Número da Versão
-//      Revisão
-//
-// É possível especificar todos os valores ou usar como padrão os Números de Build e da Revisão
-// utilizando o "*" como mostrado abaixo:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
