--- MenuAnimado1/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="MaterialDesignColors" version="1.1.2" targetFramework="net461" />
-  <package id="MaterialDesignThemes" version="2.3.1.953" targetFramework="net461" />
-</packages>
