+++ MenuAnimation.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27004.2009
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MenuAnimation", "MenuAnimation\MenuAnimation.csproj", "{F5B2AEAA-8BA7-4E14-846B-D8610EA6C9B2}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{F5B2AEAA-8BA7-4E14-846B-D8610EA6C9B2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F5B2AEAA-8BA7-4E14-846B-D8610EA6C9B2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F5B2AEAA-8BA7-4E14-846B-D8610EA6C9B2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F5B2AEAA-8BA7-4E14-846B-D8610EA6C9B2}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {482795C4-7094-4C24-BC49-ED408DC43E4D}
+	EndGlobalSection
+EndGlobal
