+++ MenuAnimation/App.xaml.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Configuration;
+using System.Data;
+using System.Linq;
+using System.Threading.Tasks;
+using System.Windows;
+
+namespace MenuAnimado1
+{
+    /// <summary>
+    /// Interação lógica para App.xaml
+    /// </summary>
+    public partial class App : Application
+    {
+    }
+}
